BenchmarkDotNet is not compatible with xunit.v3 new hosting. So move to own project
