+++ NET/Mbc.Ads.Mapper.Test/Mbc.Ads.Mapper.Test.csproj
@@ -8,8 +8,7 @@
     <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
 
     <OutputType>Exe</OutputType>
-
-    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+    <UseAppHost>true</UseAppHost>
   </PropertyGroup>
   <PropertyGroup>
         <NoWarn>NU1702;CS7022</NoWarn>
@@ -18,7 +17,6 @@
     <ProjectReference Include="..\Mbc.Ads.Mapper\Mbc.Ads.Mapper.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.15.8" />
     <PackageReference Include="coverlet.collector" Version="10.*">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
