+++ NET/Mbc.Benchmark.Test/BenchmarkMain.cs
@@ -1,6 +1,8 @@
 ﻿using BenchmarkDotNet.Running;
-using Mbc.Ads.Mapper.Test;
+using Mbc.Ads.Mapper;
+using Mbc.Pcs.Net.State;
 
 /* Works only without debugging in release configuration. */
 
+BenchmarkRunner.Run<PlcHeartBeatTimerBenchmark>();
 BenchmarkRunner.Run<MarshallingPerformanceTest>();
