--- NET/Mbc.Pcs.Net.Test/BenchmarkMain.cs
@@ -1,6 +1,0 @@
-﻿using BenchmarkDotNet.Running;
-using Mbc.Pcs.Net.Test.State;
-
-/* Works only without debugging in release configuration. */
-
-BenchmarkRunner.Run<PlcHeartBeatTimerBenchmark>();
