+++ NET/Mbc.Pcs.Net.Test/Mbc.Pcs.Net.Test.csproj
@@ -7,12 +7,12 @@
     <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
 
     <OutputType>Exe</OutputType>
+    <UseAppHost>true</UseAppHost>
   </PropertyGroup>
   <PropertyGroup>
-        <NoWarn>NU1702;CS7022</NoWarn>
+    <NoWarn>NU1702;CS7022</NoWarn>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.15.8" />
     <PackageReference Include="coverlet.collector" Version="10.*">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
