+++ NET/Mbc.Pcs.Net/Mbc.Pcs.Net.csproj
@@ -3,11 +3,11 @@
     <!-- common nuget configuration see Directory.Build.props-->
     <TargetFrameworks>net10.0;net8.0;netstandard2.0</TargetFrameworks>
     <Version>5.2.0</Version>
-    
+
     <Product>Mbc.Pcs.Net</Product>
   </PropertyGroup>
   <PropertyGroup>
-        <NoWarn>NU1702</NoWarn>
+    <NoWarn>NU1702</NoWarn>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\Mbc.Ads.Mapper\Mbc.Ads.Mapper.csproj" />
